recipe card css half-done. 8 x random cards being generated. config of api call separated as much as possible from logic. new logic file created.
