+++ views/partials/recipes-goal-bar.handlebars
@@ -1,5 +1,5 @@
 <!--goalBarhandlebars-->
-{{!-- <div class="goalBar">
+<div class="goalBar">
   <div class="row">
     <div data-sticky-container>
       <div class="title-bar" data-sticky data-options="marginTop:0;" style="width:100%" data-top-anchor="1"
@@ -9,21 +9,6 @@
     </div>
   </div>
 </div>
-<div class="row">
-  <div data-sticky-container>
-    <div class="title-bar" data-sticky data-options="marginTop:0;" style="width:100%" data-top-anchor="1"
-      data-btm-anchor="content:bottom">
-      <div class="column small-6 colBar"><i class="fas fa-utensils fa-lg bari"></i>Ideal Calorie Intake:
-        {{user.idealCalories}} kcal</div>
-        <div class="column small-6 colBar"><i class="fas fa-running fa-lg bari"></i>Activity: {{user.activity}}</div>
-        <div class="column small-6 colBar"><i class="fas fa-glass-whiskey fa-lg bari"></i>Water Intake:
-          8 glasses</div>
-        <div class="column small-6 colBar"><i class="fas fa-bed fa-lg bari"></i>Sleep: 8 Hours</div>
-        <div class="column small-6 colBar"><i class="fas fa-child fa-lg bari"></i>Ideal Weight: {{user.idealWeight}} kg</div>
-      </div>
-    </div>
-  </div>
-</div> --}}
 
 <!--Display Goalbard for screensizes screens screens equal to or greater than 1200px wide-->
 <div class="row d-none d-lg-block d-xl-block">
